requirements.txt
apptools==5.2.1
bleach==5.0.1
cloudpickle==3.0.0
cmake==3.28.1
commonmark==0.9.1
configobj==5.0.8
contourpy==1.0.6
cycler==0.11.0
daal4py==2021.6.0
dask==2024.2.1
distlib==0.3.6
docopt==0.6.2
docutils==0.19
envisage==7.0.3
fairscale==0.4.13
filelock==3.8.0
fire==0.5.0
fonttools==4.38.0
fsspec==2023.12.2
geographiclib==2.0
geopy==2.4.1
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-tools==1.49.1
h5py==3.10.0
hanlp==2.1.0b45
hanlp-common==0.0.19
hanlp-downloader==0.0.25
hanlp-trie==0.0.5
hf_transfer==0.1.6
huggingface-hub==0.21.4
interchange==2021.0.4
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.3
keyring==23.11.0
kiwisolver==1.4.4
lit==17.0.6
locket==1.0.0
lxml==5.0.0
MarkupSafe==2.1.4
matplot==0.1.9
matplotlib==3.6.2
mayavi==4.8.1
mkl-fft==1.3.1
mkl-service==2.4.0
monotonic==1.6
more-itertools==9.0.0
mpmath==1.3.0
networkx==3.2.1
nibabel==5.2.1
nltk==3.8
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
pandas==1.5.0
pansi==2020.7.3
partd==1.4.1
patsy==0.5.6
pdf2docx==0.5.7
phrasetree==0.0.8
Pillow==9.3.0
pipreqs==0.4.11
pkginfo==1.9.2
platformdirs==2.5.2
pooch==1.8.1
protobuf==4.21.6
py2neo==2021.2.3
pybind11==2.10.1
pyface==8.0.0
Pygments==2.13.0
pyloco==0.0.139
PyMuPDF==1.22.5
PyMuPDFb==1.23.9
pynvml==11.4.1
pysurfer==0.10.0
python-docx==1.0.1
pytz==2022.2.1
PyYAML==6.0
readme-renderer==37.3
regex==2022.10.31
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
scikit-learn-intelex==2021.20221004.171507
scipy==1.13.1
seaborn==0.12.1
SecretStorage==3.3.3
sentencepiece==0.1.97
SimpleWebSocketServer==0.1.2
statsmodels==0.14.2
sympy==1.12
termcolor==2.1.1
tokenizers==0.11.6
toolz==0.12.1
toposort==1.5
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.64.1
traits==6.4.3
traitsui==8.0.0
transformers==4.25.1
triton==2.0.0
twine==4.0.2
typing==3.7.4.3
ushlex==0.99.1
verde==1.8.0
virtualenv==20.16.5
vtk==9.3.0
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.4.2
xarray==2023.12.0
xlrd==2.0.1
xxx==0.0.1
yarg==0.1.9
